// react native/Mobile/WGFMobile/services/Model/PaymentModel.tsx
export class PaymentCheckoutResource {
    cardholderName: string = "";
    cardNumber: string = "";
    email: string = "";
    expirationMonth: number = 0;
    expirationYear: number = 0;
    Address: string = "";
    CVC: string = "";
    amount: number = 0;
    teamId: number = 0;
    userId: number = 0;
    type: string = "";
}