// react native/Mobile/WGFMobile/services/Model/TeamModel.tsx
export class TeamModel {
    teamId: number = 0;
    teamName: string = "";
    defaultTeam: string = "";
    teamRole: number = 0;
    productLevel: number = 0;
    addOnCamp: number = 0;
    addOnRoku: number = 0;
    teamStatus: string = "";
    teamGuid: string = "";
    teamColor: string = "";
};

export class TeamUserSetting {

    teamId: number = 0;
    userId: number = 0;
    firstName: string = "";
    lastName: string = "";
    emailAddress: string = "";
    phone: string = "";
    teamRole: number = 0;
    dateExpired: string = "";
    addUserId: number = 0;
    fullName: string = "";
    userPassword: string = "";

}